feat: add routes privates

+++ api-spacetime/src/routes/memories.ts
@@ -4,8 +4,15 @@ import { z } from "zod"
 import { prisma } from "../lib/prisma"
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/api/memories", async () => {
+  app.addHook("preHandler", async (request) => {
+    await request.jwtVerify()
+  })
+
+  app.get("/api/memories", async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createAt: "asc",
       },
@@ -20,7 +27,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get("/api/memories/:id", async (request) => {
+  app.get("/api/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -32,6 +39,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     })
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     return memory
   })
@@ -50,14 +61,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "",
+        userId: request.user.sub,
       },
     })
 
     return memory
   })
 
-  app.put("/api/memories", async (request) => {
+  app.put("/api/memories", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -71,7 +82,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
     const { id } = paramsSchema.parse(request.params)
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -85,12 +106,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.delete("/api/memories/:id", async (request) => {
+  app.delete("/api/memories/:id", async (request, reply) => {
     const paramsShema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsShema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
